checkout fixed
